MAINT: Remember that usize is used for len()

+++ src/error.rs
@@ -2,6 +2,6 @@
 pub enum ParseError {
     IncompleteHeader,
     IncompleteFrame,
-    InvalidVectorLength { expected: u64, found: u64 },
+    InvalidVectorLength { expected: usize, found: usize },
     InvalidNumberFormat(String),
 }
